energy-in-schools/pyproject.toml
[tool.poetry]
name = "samsung_school"
version = "0.4.0"
description = "Energy in Schools Project energyinschools.co.uk"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^2.1"
aldjemy = "=0.8.0"
celery = "^4.2.1"
djangorestframework = "^3.8.2"
funcy = "^1.10.3"
redis = "3.2.1"
SQLAlchemy-UTCDateTime = "^1.0.4"
SQLAlchemy-Utils = "^0.33.6"
django-configurations = "^2.1"
django-cors-headers = "^2.4.0"
djangorestframework-simplejwt = "^3.2.3"
drf-yasg = "^1.9.2"
django-encrypted-model-fields = "^0.5.8"
django-taggit = "^0.23.0"
django-taggit-serializer = "^0.1.7"
django-secretballot = "^1.0.0"
django-private-storage = "^2.1.3"
django-storages = "^1.7.1"
django-tinymce = "^2.7.0"
psycopg2-binary = "2.8.6"
stringcase = "^1.2.0"
django-enumfields = "^1.0.0"
paho-mqtt = "^1.4.0"
singleton-decorator = "^1.0.0"
colour = "^0.1.5"
pyowm = "^2.9.0"
djangorestframework-csv = "^2.1.0"
django-cacheops = "^4.1"
drf-serializer-cache = "^0.3.3"
feedparser = "^5.2.1"
httpsig = "^1.2.0"
pip = "^18.1"
django-safedelete = "^1.0.0"
django-filter = "^2.0"
XlsxWriter = "^3.0.1"
flex = "^6.13.2"
swagger-spec-validator = "^2.3.1"
makepass = "^0.11.0"
martor = "^1.4"
django-cleanup = "^4.0"
cryptography = "2.4.2"
pyjwt = "1.7.1"
python-dateutil = "^2.8"
sqlalchemy = "1.2.15"

[tool.poetry.dev-dependencies]
tblib = "^1.3"
responses = "^0.10.5"
factory-boy = "=2.11.1"
coverage = "^4.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
